Extract result display into extra class
